Consolidate deploy execution
Consolidation the two different deploy functions into a single
function.  It was a bit tricky because their failure semantics are
different when goals are not a match, but once you realize the mode
will only be set on side-effect goals, you can sort it out.

Convert side-effect mode values to an enum type.