Remove use of goal fulfillment callback
Move away from deprecated goal fulfillment callback to a two-phase
goal execution.  The old callback is now the executor for all
invocations of the Kubernetes deploy goal.  This leaves the goal,
which no longer has a side-effect fulfillment, in the "in_process"
state.  The event handler now looks for this state and completes all
invocations of this goal.