Remove environment, namespaces, & default ingress
Since any unique instance of an SDM using this pack must be registered
with a unique name, there is no need to differentiate instances using
an environment property.  Remove the environment property from the
application data and its significance from the goal details.

Similarly, it was not helpful to have namespaces configuration
property on the SDM since the name of the SDM is unique.  Requesting
deployments of an SDM not managing the applications namespace would
result in a goal that was never even attempted.  Now, it will be
attempted and RBAC will control whether it is successful or not.

Remove context from SDM pack options, just use the current context if
using the default Kubernetes credentials.

It only makes sense to provide a default ingress if deploying locally.
Previously a default ingress was generated for all deployments when
the SDM was in local mode.  Remove the default ingress logic from this
package and add correct implementation to k8s-sdm.

Refactor creation of the goal result externalUrls to a function.