Remove fulfillment property on goal, just use name
Since the fulfillment was just used to set the name of the fulfillment
and the name property of the FulfillmentRegistration was not used but
required, just use the name property as the name of the fulfillment
instead of introducing a new property.

Unfortunately, the name property is required on
FulfillmentRegistration but we can supply a default.  I do not know of
a way to provide a default value for one property in an interface.