Add sync mode PHASE20
Make Kubernetes generic API client smarter about names and namespaces.
This hard codes information about resources and operations that
require resources, which may not be a good idea but it is the only way
I see to provide good feedback instead of the error "the server does
not allow this method on the requested resource" back from the API.

Send proper header for patch operations.