Clean up SyncRepoRef interface
Technically this is breaking but in reality the old interface was
broken.  The properties that were removed were not used and not
useful.  Adding providerId and using it in selection if it is present
provides the necessary functionality.  Not extending SyncRepoRef from
RepoRef removes the broken idea of providing a SHA but loses the
perhaps useful providing a path.  The latter can still be done by
providing a RemoteRepoRef directly.

[changelog:fixed]