Account for partial specs changing resource name
It was possible for the supplied partial specs in the
KubernetesApplication object the change the name and namespace of the
resource.  Changing the name makes sense.  Changing the namespace does
not seem like a good idea.  Account for the name possibly being
different and ensure that the namespace is not changed.

This change required reworking the way resources are deleted.
Previously, other than secrets, they were deleted by name.  Now, like
secrets, they are all deleted by label matching.  Add tests for
application deletion.  This change should maintain backward
compatibility, although the delete functionality is not widely used.

Closes #81