Set KubernetesDeploy goal back to requested
The new FulfillGoalOnRequested mechanism works when the goal, as the
name suggests, is in the requested state, not the in_process state.
Since we still have the old event handler for the in_process state,
SDMs using this pack should be able to fulfill both the old and new
style goals.